+++ packages/yourloops/test/unit/lib/patient/patient.hook.test.tsx
@@ -89,7 +89,6 @@ describe('Patient hook', () => {
         filters: {
           pendingEnabled: false,
           manualFlagEnabled: false,
-          telemonitoredEnabled: false,
           timeOutOfTargetEnabled: false,
           hypoglycemiaEnabled: false,
           dataNotTransferredEnabled: false,